duui-open-webui/requirements.txt
opencv-python
torch==2.8.0
torchaudio==2.8.0
setfit==1.0.3
sentencepiece==0.2.0
protobuf
germansentiment==1.1.0
pytorch-lightning==2.4.0
numpy==1.26.3
fastapi==0.115.12
dkpro-cassis==0.9.1
uvicorn[standard]==0.27.1
pydantic-settings==2.0.2
torchmetrics==1.2.0
six==1.16.0
pandas==1.4.3
pysentimiento==0.7.3
starlette==0.40.0
transformers
accelerate==1.3.0
soundfile==0.13.1
pillow==11.1.0
scipy==1.15.2
backoff==2.2.1
peft==0.13.2
qwen-omni-utils[decord]
qwen-vl-utils[decord]==0.0.8
huggingface_hub[hf_xet]
 python-multipart